separate the keybindings function into its own module
